Fix issue where rebuild on --watch was causing an ENOENT error
This was very hard to debug!

Basically the first build would add a .client to loaderConfig.paths in
aurelia-cli/lib/build/bundle.js on line 211.  This would allow
Aurelia.js to find client.js correctly.  But on the rebuild with —watch
require js would then recombine that path with the original and get a
weird path such as adam/client/assets/client rather than
adam/dscsa/node_modules/client which would result in a ENOENT error.
For some reason this configuration resolves this error.