Turn of sourcemaps
Right now the bundles name is assets/client.js this is then used by
Aurelia’s aurelia-cli/lib/build.js line 128 and line 153 to set the
source map url.  However since that name is relative and its already in
the assets folder, the browser looks for the map file in
clients/assets/assets which gets a 404 and logs loading errors on the
sever.  I am sure if I played around with it I could get it to work but
don’t care enough right now to properly fix